requirements.txt
qdrant-client
langchain
langchain-community
langchain-core
langchain-google-genai
langgraph
sentence-transformers
streamlit
python-dotenv
pillow
fastembed